Refactored tests and websocket/rpc clients DI/setup

+++ SubstrateClientSwift/Classes/RPC/RpcRequest.swift
@@ -3,7 +3,7 @@ import Foundation
 /// RPC request with generic params
 struct RpcRequest<T: Codable>: Codable {
     var jsonrpc = "2.0"
-    let id: Int64
+    let id: Int32
     let method: String
     var params: T
 }
